Fix Candidate.addAssessments()
1. When an image with a description page that already contains
the `{{Assessments}}` template (e.g. because the image is featured
on some Wikipedia) is promoted to FP status, the bot always fails
to add the parameter `featured=1` to the `{{Assessments}}` template.
That’s astonishing because the method `Candidate.addAssessments()`
contains code which deals with exactly that case.
Just reducing the indentation level of the last line `self.commit(...)`
fixes this.  That line must be applied regardless whether
there is already an `{{Assessments}}` template or not,
else the prepared changes are just not committed.

2. After fixing the first bug another one becomes obvious.
The method contains two lines which should remove already existing
`featured=...` or `com-nom=...` parameters from the `{{Assessments}}`
template, but they do not work if the parameter is the first one
in the template.  Why?  Both code lines assume that the parameter
is preceded by a `|`, but the regex which extracts the parameters
does not capture the `|` before the first parameter.
We can fix this simply by moving the `\|` into the capturing
parentheses of the regex, i.e., replacing `\|(.*)` by `(\|.*)`.

3. While we are at it we can also make that regex more foolproof.
I have seen people putting several templates (even completely
unrelated ones) on a single line.  If somebody would put a template
on the same line right after the `{{Assessments}}` template,
the `.*` would cheerfully capture not only the parameters,
but also the `}}` and the complete next template until its `}}`.
OK, this is a rather unlikely problem, but I have seen people
creating such Wikitext lines, and we can simply handle this by
just making the greedy `*` quantifier non-greedy: `(\|.*?)` fixes this.

4. To be on the save side let’s properly escape the leading `{{`
and trailing `}}` in the regex.  Yes, the regex works as it is,
but normally every literal `{` and `}` in a regex should be escaped,
and currently it depends on the regex engine and maybe
even on its version whether the regex works without the escapes or not.