Simplify Candidate.addAssessments()
While we are at this method, let’s tweak and simplify
some minor points in the code.

1. We do not need to create an empty list first and then append
an entry to it, we can just define a list with that single entry.

2. Both people reading the code and linters/type checkers may stumble
over the double use of the local variable `params`; unlike its name
suggests, it is first used for the complete result of the `re.search()`
call which is either a `re.Match` object or `None`.
Using two variables, `match` and `params`, makes the code clearer
and avoids any misunderstandings.

3. While the previous change may appear pedantic, it has the advantage
that it allows us to drop the repeated regex search: we can simply
use the position values `match.start(0)` and `match.end(0)`
from the first search to construct the new text with the new version
of the `{{Assessments}}` template.

4. We can apply the first TODO comment and remove the support
for the old `com=…` parameter: a search shows that not a single
featured picture still contains the `com=…` parameter.
(But we must not remove the support for the `subpage=…` parameter:
a search shows there are still ~ 1.871 featured pictures
which use that variant of the parameter.)

5. Instead of searching for the first occurrence of `|` in `params`
and checking if its position is `!= 0`, we can just test
`if params[0] != "|"`.

6. The commented line `# new_text = re.sub(r'({{\s*[Ii]nformation)',...`
can and should be removed.  The existing code above it
which calls `findEndOfTemplate()` works well and is much better,
so the commented code is just confusing and should go.