Fix and improve Candidate.setFiles()
The method `Candidate.setFiles()` tries to extract
the filenames of all nominated images from a set nomination.

Currently this method is rather picky and works only well
if all filenames are listed with the `File:` prefix,
each filename appears on its own line, and not even
simple spaces are allowed in front of the `File:` prefix.
But from normal gallery pages users are accustomed
to omit the `File:` prefix, to use the old `Image:` prefix
instead of it, and/or to put some spaces as indentation
before the individual filenames.  Therefore the method
often fails to handle lax set nominations, and some
experienced users must handle these nominations manually.
That’s cumbersome.

We can make the method a bit more tolerant and teach it
to handle also the `Image:` prefix and/or leading spaces;
and, if this does not help, to search for entries
which look with sufficient probability like filenames
thanks to a common image file extension.

This commit also fixes two bugs in the old implementation:

1. If a set nomination does not contain a `<gallery>` element
with is formatted exactly as expected by the very
restrictive regex in the old code, the old implementation
raises an `AttributeError` exception and the bot crashes
because the old code does not check whether the result of
    `re.search(r"<gallery([^\]]*)</gallery>", ...)`
is a valid `re.match` object or just `None`.
This happens e.g. with the following nomination:
    [[Commons:Featured picture candidates/Set/
    Berlin Neue Nationalgalerie]]
just because the captions of the images contain links,
but the regex in the old code does not like this.
The new implementation handles this case correctly.

2. If a set nomination has been renamed and the old page
from the list of nominations contains just a redirect,
the old code does not follow that redirect; instead
it searches for the `<gallery>` element in the text
of the redirect page, this triggers again
the first bug and stops the bot.
This can be observed e.g. with this nomination:
    [[Commons:Featured picture candidates/Set/
    Callionima falcifera mounted specimen]]
The new implementation follows the redirect and returns
the correct filenames; and if this does not work,
it prints a (hopefully) useful error message and allows
the script to continue.

Finally the new implementation makes the results of the method
more uniform and reliable; it guarantees that every filename
in the returned list starts with the proper `File:` prefix
and does not contain any irritating spaces at the beginning,
at the end or between prefix and filename.

I have tested and compared the results of old and new code
with all set nominations between January 2022 and April 2025.
Of 111 set nominations, 12 show differences between
old and new implementation.  Four nominations crash
the old code (see above), the new one works in all cases;
in two nominations the old code does not find
a single image, the new code finds all images of these sets;
in 6 nominations the old results contain stray spaces
or underscores, the new ones are clean.
The new code did not introduce new errors.