Fix and rewrite Candidate.addToGalleryPage()
The method `Candidate.addToGalleryPage()` causes many problems
which are best solved by a complete rewrite.  For a description
of the problems and guidelines for the rewrite, please see

https://commons.wikimedia.org/wiki/User:Aristeas/FPCBot/Rewriting_Candidate.addToGalleryPage()

Just in case that this page ever gets damaged or deleted,
here is a permalink to its current state:

https://commons.wikimedia.org/w/index.php?title=User:Aristeas/FPCBot/Rewriting_Candidate.addToGalleryPage()&oldid=1038498025

The new code has been tested thoroughly with the gallery links
from all nominations between January 2024 and April 2025,
comparing the results with the results of the old code.