Use unified method to get a clean set title
In various places we need the clean title of a set nomination,
i.e. without the ‘Commons:Featured picture candidates/’ part
and most often (but not always) also without the ‘Set/’ part.
For this purpose, let’s add a simple unified method
`Candidate.cleanSetTitle()` and use it in all places.