Improve decision about ‘com-nom’/‘subpage’ parameter
The methods `Candidate.addAssessment()` and `notifyUploader()`
have to decide whether we need the ‘com-nom’ resp. ‘subpage’
parameter for the template or not, and which value it has.
This allows for misunderstandings because there are different
reasons to add the parameter which are not handled together.
Let’s make the code clear and unambiguous.