Simplify and improve wikipattern()
The function `wikipattern()` is used to prepare a short string,
normally a filename, page name or nomination subpage name,
so that it can be used in a regex search string:
(i) it escapes characters which have a special meaning in a regex;
(ii) it replaces both spaces and underscores by `[ _]`
because the MediaWiki software handles them as interchangeable
in filenames, page names, etc.

We can replace the body of this function by a single line.
It just relies on the function `re.sub()` which knows best
which characters should be escaped for a regex that will
be handled by the regex implementation it belongs to.
This also makes the function much more reliable
and easier to maintain, because it fixes several issues:

1. On the one hand the old code escapes some characters,
e.g. `=` and `-`, which have a special meaning only
in special contexts in a regex and therefore do not need
to be escaped.  On the other hand it does not escape
important characters like `|` which can very well appear
in Wikitext fragments; it does not even escape `.`
which appears in many filenames and *must* be escaped.
If a `.` is not escaped, it matches *any* character;
if a `|` is not escaped, it means an alternative;
depending on the purpose to which the `wikipattern()`
result is used this could very well result in false-positive
search and test results.  Call this a bug if you want.

2. In the old code each character which is to be escaped
must be listed twice: first in the capturing character class
in the outer code, then in the long test conditions
of the nested function `rep()`.  This makes it burdensome
to maintain the old implementation: the programmer
must always remember to keep both parts synchronous,
the character class and the test conditions must handle
exactly the same list of characters.

3. This is even related to a sleeping bug in the old code
which would wake up immediately if we forget to keep
the two parts perfectly synchronous.  The nested function
`rep()` misses a final `else:` statement.  Therefore
it has an implicit default return value of `None`
which would be returned if a character passed to `rep()`
does not meet any of the test conditions.
This could happen very well if we add another character
to the capturing character class in the outer code,
but forget to add it to the tests in `rep()`.
Then `rep()` would just return `None`.
But if a `re.sub()` call which uses a replacement function
receives `None` from that function it just omits
the tested string, in our case the character.
So the function would mutilate the string.

The points 2 and 3 mean that the old code is rather fragile
because it works only well as long as outer and inner
character tests are perfectly balanced; and this makes
the maintenance unnecessarily difficult.

There is a single caveat for the new code.  Our search pattern
`r"(?:\\ |_)"` is based on the fact that Python’s `re.escape()`
function escapes simple spaces as `\ `.  This is necessary
only if we use the flag `re.VERBOSE`, but not in general,
so different regex implementations (and maybe even versions)
may handle this differently.  If the regex implementation
ever changes so that `re.escape()` does not escape spaces,
we just need to change our own `re.sub()` search pattern
from `r"(?:\\ |_)"` to `r"[ _]"`.  I have added a general
assertion to make sure that we notice it if the handling
of spaces by `re.escape()` changes.