Increase pywikibot.showDiff() context to 3
The `-dry` option (running the bot and seeing what changes
it would be make without applying them) and the possibility
to run the bot in step-by-step mode (without the `-auto` option,
to check each change it proposes) are wonderful test tools.
It is essential that they show a diff for every upcoming change.

But if one wants to review a change the bot is going to make
it is often not enough to see the line(s) it wants to add or
remove; one can assess the line(s) much better if one also sees
the context of the change, i.e. *where* on the affected page
the bot wants to add or remove the line(s).  Especially when
adding a new entry to a gallery page, or rotating the images
on the front-page list of recently promoted featured pictures,
the context is crucial; some of the most difficult problems
with FPCBot is that it often puts new FPs into the wrong
section on a gallery page.

In order to review changes more easily and accurately,
I have set the `context` parameter for `pywikibot.showDiff()`
to `3`, i.e. 3 lines of context before and after each change.
This has helped me a lot.  Therefore I propose to use
3 lines of context for the diff in dry or step-by-step mode,
while keeping the old value of 0 (no context) if the bot
runs with the `-auto` but without the `-dry` flag --
this keeps the log files short, and if you want to dig
into a problem and need more context, you will run the bot
with the `-dry` option and/or without the `-auto` option.