Add missing NL at beginning of <gallery> element
When notifying the nominator that a set nomination was successful,
we use a `<gallery>...</gallery>` element to show all images of the set.
But there is no newline character (NL) between the opening `<gallery>`
and the first entry.  That’s uncommon and makes it hard to read
and edit the Wikitext.  So let’s add the missing NL.