Remove converttostr() function
The `converttostr()` function, called just a single time, looks useful,
but in the end it is just a detour to the common `str.join()` method.
The latter is widespread and easy to understand, and IMHO the code
gets even clearer if we just use it directly.  So let’s do that.