Update deprecated date/time functions
The Python Standard Library method `datetime.datetime.utcnow()`
is deprecated and the use of `datetime.datetime.now()`
(which does not explicitly specify the timezone) is ticklish.
So we must update these methods to the recommended ones,
namely to the offset-aware `datetime.datetime.now(datetime.UTC)`.

Furthermore, the Pywikibot 10.x documentation recommends to use
`BasePage.oldest_revision.timestamp` to get the creation time and
`BasePage.latest_revision.timestamp` to get the last edit time
of any page.  The method `BasePage.editTime()` is deprecated,
searching `BasePage.revisions` and then getting the timestamp
is a bit cumbersome and discouraged.

This commit updates the code to use the recommended and
up-to-date functions resp. properties.

There was one difficulty: the Pywikibot `timestamp` attributes
yield `pywikibot.Timestamp` objects which are in UTC
(as every MediaWiki timestamp) but still use an offset-naive format.
Therefore they cannot be compared to offset-aware `datetime` objects
which are highly recommended both by the Python documentation
and by the Pywikibot documentation and avoid any miscalculations
due to confusion with local times.  Therefore we convert all
`pywikibot.Timestamp` objects right away to offset-aware
`datetime` objects, so the comparison is both easy and correct.
The conversion via `timestamp.replace(tzinfo=datetime.UTC)` relies
on the fact that all MediaWiki timestamps are actually in UTC.

I have compared the results of the old and the new code,
especially for the crucial `delta` values (time since creation
or since the last edit of a nomination), and they are identical.
The comparison was made on a computer with a timezone different
from UTC, so the correspondence between old and new code
should not be pure coincidence, but prove the correctness
of the updated code.