Don’t refer to title if reading timestamp has failed
But when we think again about it, it’s probably not a good idea
to include the title of the nomination in the error message
if querying the timestamp of the last revision has raised
an exception -- then `self.page.title()` will probably raise
yet another exception and we stop with an unhandled exception
instead of printing a helpful error message.  Therefore remove
the reference to `self.page.title()` from the message.

Returning the surrogate date/time directly instead of saving it
into `self._creationTime` first (because it is just a surrogate)
allows us to drop the `else: ...` clause from the
`try: … except: …` construct.