Improve method Candidate.isSet()
1. Make the docstring more precise: we do not search the nomination
subpage, i.e. its contents, for the magical `/[Ss]et/` substring,
but the name of the nomination subpage.

2. Require a `/` after the `[Ss]et`, too -- without it the subpage
could also be a sloppy file nomination whose name begins with ‘Set’
and misses the obligatory ‘File:’ prefix.

3. Tolerate spaces before/after `[Ss]et` -- they are wrong,
of course, but people sometimes add odd spaces to the names
of nomination subpages and this function should also find
set nominations with imperfect names.

4. No need for four lines with a complete `if ...: ... else: ...`,
we just need to test `is not None`.