Improve cutTitle() for set nominations
The method `Candidate.cutTitle()` is used to cut and align the title
of a nomination to a fixed width for nice tabular output.
But with set nominations it keeps the complete namespace etc.:
‘Commons:Featured picture candidates/Set/...’ makes the name
of a set nomination almost unreadable if cut to 50 chars.
Let’s get rid of the namespace etc. and start the output
right at the ‘Set/’ prefix, this makes the output of `-info` etc.
much clearer and easier to understand.