Add docstring and structural comments to main()
The most important function of a module should certainly have
its own docstring, and in this case it is especially necessary
to explain the intended use of the `*args` argument
(although currently it does not really work as advertised,
but that will be fixed with the next commit).

The structural comments are useful to avoid misunderstandings
and help to prepare the next commit which will tidy up
and improve the structure of this code.  Then the comments
will also serve as landmarks for any diff tools which show
the changes in that commit ;–).