Improve handling of CLI arguments in main()
The handling of the various CLI arguments in `main()` does not
work well.  Apparently some things got mixed up during the
continued development of the script and therefore interfere
with each other.  For example, the `*args` argument seems
to be intended to allow for overriding the CLI arguments,
e.g. for test purposes, but `*args` is only used by the call
to `pywikibot.handle_args(*args)` and not taken into account
by the first block of code which handles the local options.
The direct manipulation of `sys.argv` in that block
just does not work that way: local options are never removed
(as intended) from `sys.argv`, therefore they pop up again
in the test below which tries to catch unknown arguments,
forcing us to add every single local option to the list there
(which should be unnecessary).  Because we cannot add
arbitrary strings to that list, the pattern string following
the `-match` option always triggers the ‘unknown argument’
warning, making it impossible to use the `-match` option.
And there are more and minor issues.

We can get rid of these problems by bringing the various steps
into a meaningful order.  This means that we must start with
the call to `pywikibot.handle_args()`; it allows Pywikibot
to acquire the CLI arguments for us (so we do not need to cope
anymore with `sys.argv`) and also allows `*args`, if specified,
to be used instead of the CLI arguments.  This way Pywikibot
handles all so-called global options and returns a simple list
of the remaining local options which is ready to use for us.
Then we can handle all local options in the original meaning
of that word (including the fancy `-match` argument)
which serve as modifiers for the tasks of our script.
The remaining arguments must designate the desired tasks,
so it’s easy to check for unknown arguments and then to perform
the desired tasks in the desired order.

We can also improve some minor points, e.g. the error messages
(these are really errors, as they stop our script completely,
so messages should begin with ‘Error’, not just with ‘Warning’),
and also print a clear and hopefully helpful error message
if the connection to the Commons server has failed.
(We cannot put all the countless calls of functions and methods
which require a valid connection into `try: ... except: ...`,
but it makes sense to do this here with the very first command
that connects to Commons.  So we can exit with a clear message
instead of with the stack trace that suggests a much more severe
problem than just a failed connection).