Refer to superclass with super(), document __init__()
In the code of a class we can refer to its superclass with
`super()`.  This has naby advantages (see Raymond Hettinger’s
well-known essay ‘Python’s super() considered super!’ [1]);
e.g. (to mention one of the most modest ones) that the code
continues to work seamlessly if the superclass is renamed.

We can omit `self` from the list of parameters when calling
`super().__init__()`, Python 3 just handles this for us.

Improve the docstrings for the `__init__()` methods (initializers)
of the three candidate classes in order to provide at least
the basic information.

[1] https://rhettinger.wordpress.com/2011/05/26/super-considered-super/