Make commit() a module function
The method `Candidate.commit()` is a static method, i.e. it depends
not at all on the instance it is called on (`self`), but works
like an independent function.  In many cases it is useful to make
a function a static method because it is connected closely
to the concepts and implementation of a certain class.
But in this cases there is no necessary connection: the bot program
changes also pages which are not part of the nominations
(e.g. user talk pages) and may very well need to change more pages.

Indeed one of the next commits needs to call `commit()` from
another module function.  To allow this, and also to make explicit
that `commit()` is a general function and not limited to the contents
of a candidate object, let us make `commit()` a module function.

NB: Code and comments of `commit()` are not changed in this commit,
the function is just moved up to module level.