Handle candidates chronologically, resolve redirects
The function `findCandidates()` returns a list with candidate objects
for all nomination subpages, either from the page with the current
candidates or from a log page with closed nominations.
Currently the returned list is in random order and that random order
even changes which every call to the function.  The reason is
that the function builds on `pywikibot.BasePage.templates()`
which is just intended to extract all used templates and doesn’t care
for their original on the page.

This is very irritating if one tests improvements in the bot code
and wants to compare before/after results -- the random order
of the candidate objects makes this comparison really difficult.
But it is also detrimental to the everyday duties of the bot.
Nominators are sometimes surprised why the bot closes or parks
nominations in random order; obviously humans expect the bot
to handle the nominations in chronological order.  And this gets
a noteworthy problem if the bot stops to work because it cannot
handle a certain nomination.  Right now it is hard to say *which*
nomination is the problem: it could be any nomination on the list,
so the FPC maintainers need to do investigatory work.
If the bot would handle nominations in chronological order,
it would be clear that the oldest nomination on the list which
needs to be closed or parked, but is not handled, must irritate
the bot and needs the attention of the FPC maintainers.

Therefore let’s rewrite `findCandidates()` and make it always return
the nomination subpages in the original order from the page.
Then the calling function `checkCandidates()` can reverse that list
if the page uses a descending order (i.e., newest entries first,
like the list with current nominations) or just keep the order
if the page uses an ascending order (starts with the oldest entries,
as the log pages do).

We omit damaged links (pointing to non-existent nomination pages);
it makes no sense to create a candidate object and to call its
methods when it is already clear that they must fail because
the nomination subpage does not exist.

We can and must combine this easily with another useful change.
Sometimes nomination subpages are renamed, leaving a redirect
which is still linked from the list of nominations.
This can cause problems in various parts of the bot program.
Most times it did work without noticeable problems, of course,
just because `pywikibot.BasePage.templates()` returned each
of these pages two times: one Pywikibot `Page` object for the
redirect page and one for the real nomination subpage.
That the `Page` objects for the redirect pages did not break
the `Candidate` methods was more or less pure coincidence.
So `-info` reported all these pages two times, claiming that
a redirect page contains 0/0/0 votes and is ‘Not featured’
(this made the nice table printed by `-info` rather confusing).
The task `-park` said ‘ignoring, no verified results’.
But `-close` actually tried to close redirect pages just like
real nominations, adding a ‘Not featured’ result.
This is why some old redirect pages for renamed nominations
contain an orphaned `{{FPC-results-unreviewed}}` template ;–).

OK, we can and must to better.  Let’s replace the redirects once
and for all by the true subpages, so we must not take care
for renamed nomination subpages in all the code which handles
the nominations.

Because of the checks whether each nomination subpage exists
and whether it is a redirect or not the new code is a bit slower
than the old implementation.  That’s unavoidable.  But the delay
is unimportant when compared to the time the bot needs to process
the candidates, and getting rid of actual and potential problems
with damaged links and with redirects in nomination subpages
is much more important than a short delay.