Apply '-match' filtering right in findCandidates()
Until now we have first collected all candidates in `findCandidates()`
and then filtered them with the `-match` argument (if specified)
in `checkCandidates()`.  But it is easier to filter the names
of the nomination subpages right in `findCandidates()` where we
already check the names against several criteria.
And it makes no sense to create a `pywikibot.Page` object
and to query whether it is a redirect or not, plus to create
a `FPCandidate` or `DelistCandidate` object for a candidate
which is filtered out after that.  So let’s change this.

We keep the traditional straightforward filtering approach
(namely just checking whether the `-match` argument occurs
in the name of the nomination subpage, without considering case).
But we keep the suffix in the comparison, thus allowing
the `-match` argument e.g. to filter just for specific file types.
And we can easily change this to regex matching in the future.