Replace redirects on candidate list page
Due to the recent changes we now resolve redirects to renamed
nomination subpages right when reading the page with the list
of nominations.  Therefore we do not need to take care of
redirects in nomination subpages when processing them.
But there are two exceptions:

(1) We will have to resolve the redirect again whenever we read
the list of nominations again, until the nomination is closed
and removed.  That’s not a problem, but a tiresome little waste
of time and effort.

(2) Removing the nomination subpage after closing the nomination
*is* a problem: our code delivers us the name of the renamed
nomination subpage, but because the page with the list of
nominations still links to the old name, i.e. to the redirect,
we would fail to remove that entry.

We could solve (2) by adding a complicated search for redirects
to `Candidate.moveToLog()`.  But we can do this easier and also
get rid of (1) if we resolve the redirects on the page
with the list of nominations right after finding the redirects,
i.e. at the end of the function `findCandidates()`.
This means we replace any link to a redirect by an updated link
to the corresponding renamed nomination subpage.  So starting
with the next run of the bot we will get the nomination subpage
directly instead of having to resolve the redirect again,
and human editors profit from an up-to-date list of nominations
without redirects, too.

Of course it is up to discussion whether we should also update
redirects on old log pages when running the bot with `-test`.
One could argue that old log pages should not be changed.
But resolving a redirect does not really change the content
of the log page, it just makes it a bit less complicated
by making it link directly to the nomination subpage instead of
the redirect page.  The changes are clearly documented via
the edit history of the log page, so there is no manipulation.
Furthermore, the `-test` option should be used without the
`-auto` option (or, if nevertheless, plus the `-dry` flag),
so it’s up to the user who runs the test to choose whether
a log page should be updated or not.  Finally, this affects
only relatively few instances.