Improve tests and warnings in Candidate.park()
1. The method `Candidate.park()` checks that the description page
for the nominated image exists, but does not check that the individual
images from a set nomination exist.  This is a considerable gap
because the images from a set nomination can be renamed/moved
during the nomination period.  Let’s fill the gap by testing
each image from a set nomination and breaking of if it is not found.

2. Some of the messages printed by this method are just informative,
some are warnings which describe what is wrong with the nomination
and/or the image files.  One of these warnings is emphasized
with a big ‘WARNING:’, the other ones are printed just like harmless
status information.  This is a bit confusing and easily allows
to miss an important warning.  Therefore let’s start all warnings
with a uniform ‘Warning: ’ prefix and print them using the optional
parameter `color="lightred"`, just like in other methods.

3. Until now we called `Candidate.fixHeader()` even if the success
value from the verified results is invalid, i.e., neither `yes`
nor `no`.  That’s an unnecessary detour, and because of the current
code in `Candidate.fixHeader()` that method would even append
‘not featured’/‘not delisted’ to the heading of the nomination --
that’s wrong because if the verified success value is invalid,
the nomination is *not* ‘not successful’, but needs to be fixed.
We can repair this simply by testing the success value first.

4. The docstring contained a description of the parking procedure
for successful FP candidates.  This is not necessary here anymore
(the `handlePassedCandidate()` methods of the subclasses are the
right place to explain the procedure) and even irritating as it
leaves out set and delist nominations.  Let’s replace the docstring
by an up-to-date description, explaining what *this* method does.

5. Likewise, let’s update and refine comments and variable names.

PS: Theoretically we could drop the initial test whether the
nomination subpage actually exists or not (e.g. because it has been
renamed/moved), since we test this now right in `findCandidates()`.
But this pre-condition is so fundamental that it seems better
to keep the test here, too, in order to be on the save side.
There is a tiny little chance that a nomination subpage could be
renamed right during our processing of the candidates ...